[interaction] material related into separate mod
