+++ crates/nora_object_interaction/src/lib.rs
@@ -1,13 +1,17 @@
 pub mod debug;
 pub mod event;
+pub mod material;
 mod interaction;
 
 use bevy::prelude::*;
 use nora_raycast::{RayCastPlugin, RayCastSource, RayCastSystems, RayCastable};
-use crate::event::InteractionEvent;
-use crate::interaction::{
-    DraggingGlobal, update_interactions,
-    Drag, Hover
+use crate::{
+    interaction::{
+        DraggingGlobal, update_interactions,
+        Drag, Hover
+    },
+    event::InteractionEvent,
+    material::{InteractionMaterials, set_initial_interaction_material, OriginalMaterial}
 };
 
 
@@ -39,36 +43,6 @@ impl Plugin for InteractionPlugin {
                     )
                     .with_system(set_initial_interaction_material),
             );
-    }
-}
-
-#[derive(Component, Default)]
-struct OriginalMaterial(Option<Handle<StandardMaterial>>);
-
-struct InteractionMaterials {
-    hovered: Handle<StandardMaterial>,
-    pressed: Handle<StandardMaterial>,
-}
-
-impl FromWorld for InteractionMaterials {
-    fn from_world(world: &mut World) -> Self {
-        let mut materials = world
-            .get_resource_mut::<Assets<StandardMaterial>>()
-            .unwrap();
-        Self {
-            hovered: materials.add(Color::GOLD.into()),
-            pressed: materials.add(Color::INDIGO.into()),
-        }
-    }
-}
-
-fn set_initial_interaction_material(
-    mut query: Query<(&mut OriginalMaterial, &Handle<StandardMaterial>)>,
-) {
-    for (mut original_material, material) in query.iter_mut() {
-        if original_material.0.is_none() {
-            original_material.0 = Some(material.clone());
-        }
     }
 }
